Remove two more internal crates
We actually only use those crates inside the btsieve tests and
hence, they can be inlined.
